Project: Web Page Design | METAMORPHOSIS
This repository is for a group project in designing web pages. We will be working together to design and develop a website for a specific client or purpose. The project will involve all aspects of web design, from planning and conceptualization to implementation and testing.

We will be using a variety of technologies and tools to create our website, including HTML, CSS, JavaScript, and popular web development frameworks. We will also be using Git and GitHub to manage our code and collaborate on the project.